+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/sv-parser/compare/v0.10.6...Unreleased) - ReleaseDate
 
+* [Added] get_str_trim to SyntaxTree
+
 ## [v0.10.6](https://github.com/dalance/sv-parser/compare/v0.10.5...v0.10.6) - 2021-01-28
 
 * [Fixed] constant_expression_ternary priority [#30](https://github.com/dalance/sv-parser/issues/30)